replace autotools with CMake; self-contained build
Briefly resurrecting a dorman project: I recently had a use case for
this program and found that it was a challenge to compile. Rather than
trying to wrestle with autotools, I decided to reconfigure the project
to use CMake. Additionally, the `program_options` library (and its
dependencies) from Boost version 1.57.0, which was current at the time
the code was originally written, was added as a sub-module to the main
project; so it no longer depends on any external libraries or code.